+++ test/unit/dom/within_test.dart
@@ -20,10 +20,10 @@ import 'package:react/react.dart' as react;
 import 'package:react/react_client.dart' show ReactElement;
 import 'package:react_testing_library/dom/queries.dart' show WithinQueries;
 import 'package:react_testing_library/react_testing_library.dart' as rtl;
-import 'package:react_testing_library/src/util/console_log_utils.dart';
 import 'package:test/test.dart';
 
 import '../dom/queries/shared/scoped_queries_tests.dart';
+import '../util/prints_and_logs_recording.dart';
 import '../util/rendering.dart';
 import '../util/shadow_dom.dart';
 
@@ -74,22 +74,21 @@ void main() {
         })
       ]) as ReactElement);
 
-      final logs = recordConsoleLogs(rtl.screen.debug);
-      expect(logs, equals([logs.first, logs.first]), reason: 'screen.debug() both prints and console.logs the dom');
-      expect(
-          logs.first,
-          contains(
-            '    <div>\n'
-            '      <label\n'
-            '        for="number-input"\n'
-            '      />\n'
-            '      <input\n'
-            '        id="number-input"\n'
-            '        type="number"\n'
-            '        value="3"\n'
-            '      />\n'
-            '    </div>',
-          ));
+      final printCalls = recordPrintCalls(rtl.screen.debug);
+      expect(printCalls, [
+        contains(
+          '    <div>\n'
+          '      <label\n'
+          '        for="number-input"\n'
+          '      />\n'
+          '      <input\n'
+          '        id="number-input"\n'
+          '        type="number"\n'
+          '        value="3"\n'
+          '      />\n'
+          '    </div>',
+        ),
+      ]);
     });
 
     test('exposes all the expected queries', () {