+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # React Testing Library Changelog
 
-## 1.1.5
+## 1.1.9
+*  Fix `render` swallowing errors on component mount
 
+## 1.1.8
+* (CI-only change) Add Dart 2.13.3 to CI matrix
+
+## 1.1.7
+*  Fix typing in README example
+
+## 1.1.6
+* (CI-only change) Add dependabot
+
+## 1.1.5
 * Show a valid minimum version for this library in the README.
 * Attempt to fix Dartdoc generation on pub.dev
 
 ## 1.1.4
-
 Initial public release of the library.